# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: BrickDeck Catalog API
  version: v1
  description: >
    REST API for the BrickDeck LEGO collection platform. This file is a
    hand-maintained summary of the live contract. The authoritative,
    always-current spec is generated by springdoc at runtime:
    GET /v3/api-docs (Swagger UI at /swagger-ui/index.html).
    Regenerate/refresh this file from the running API when endpoints change.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/sets:
    get:
      tags: [Catalog]
      summary: List cached sets (paginated)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
        - { name: sort, in: query, schema: { type: string, default: "externalSetNumber,asc" } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageBrickSetResponse" }
...
  /api/v1/sets/{setNumber}/missing-parts:
    get:
      tags: [Catalog]
      summary: Missing-pieces report for a target set
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: setNumber, in: path, required: true, schema: { type: string } }
        - { name: missingOnly, in: query, schema: { type: boolean, default: false } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 50 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MissingPartsReport" }
        "401": { description: Unauthorized }
        "404": { description: Set or inventory not imported }
...
  /api/v1/price-snapshots:
    post:
      tags: [Pricing]
      summary: Record a price snapshot (auth)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddPriceSnapshotRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PriceSnapshotResponse" }
        "400": { description: Invalid input }
        "401": { description: Unauthorized }
        "404": { description: Set not found or importable }
    get:
      tags: [Pricing]
      summary: List the user's price snapshots (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: setNumber, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PagePriceSnapshotResponse" }
        "401": { description: Unauthorized }
  /api/v1/price-snapshots/{id}:
    delete:
      tags: [Pricing]
      summary: Delete a price snapshot
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
        "401": { description: Unauthorized }
        "404": { description: Not found or not owned }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    BrickSetResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        externalSetNumber: { type: string }
        name: { type: string }
        yearReleased: { type: integer }
        themeId: { type: string, format: uuid }
        themeName: { type: string }
        externalThemeId: { type: integer }
        numberOfParts: { type: integer }
        imageUrl: { type: string }
        externalUrl: { type: string }
        source: { type: string }
        cacheStatus:
          type: string
          enum: [LOCAL_CACHE_HIT, IMPORTED_FROM_REBRICKABLE, EXTERNAL_SEARCH_RESULT]
...
    MissingPartLine:
      type: object
      properties:
        partNumber: { type: string }
        partName: { type: string }
        partImageUrl: { type: string }
        colorExternalId: { type: integer }
        colorName: { type: string }
        colorRgb: { type: string }
        required: { type: integer }
        owned: { type: integer }
        missing: { type: integer }
    MissingPartsReport:
      type: object
      properties:
        setNumber: { type: string }
        totalRequired: { type: integer }
        totalOwned: { type: integer }
        totalMissing: { type: integer }
        completionPercentage: { type: number, format: double }
        missingLineCount: { type: integer }
        lines:
          type: array
          items: { $ref: "#/components/schemas/MissingPartLine" }
        page: { type: integer }
        size: { type: integer }
        totalLines: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
...
    PriceCondition:
      type: string
      enum: [NEW, USED]
    PriceSource:
      type: string
      enum: [MANUAL]
...
    AddPriceSnapshotRequest:
      type: object
      required: [setNumber, amount, currency, condition, observedAt]
      properties:
        setNumber: { type: string }
        amount: { type: number }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        condition: { $ref: "#/components/schemas/PriceCondition" }
        observedAt: { type: string, format: date }
        store: { type: string, maxLength: 255 }
        url: { type: string, maxLength: 1024 }
    PriceSnapshotResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        setNumber: { type: string }
        amount: { type: number }
        currency: { type: string }
        condition: { $ref: "#/components/schemas/PriceCondition" }
        source: { $ref: "#/components/schemas/PriceSource" }
        observedAt: { type: string, format: date }
        store: { type: string }
        url: { type: string }
        createdAt: { type: string, format: date-time }
    PagePriceSnapshotResponse:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            content:
              type: array
              items: { $ref: "#/components/schemas/PriceSnapshotResponse" }
...
    PageBrickSetResponse:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            content:
              type: array
              items: { $ref: "#/components/schemas/BrickSetResponse" }
...
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }